Fix a long standing and super annoying bug in the menu where long lists would crash the program by putting a fixed window contraint in place.  Also added support for page-up, and page-down to make moving through menus easier & better.
Made a few tweaks to simplify the tools interface as the summary part is frankly no longer relevant.

Added a "summarize_text" tool as it occurred to me that that might be a useful tool to have (haven't seen it actually get _USED_ yet, but...)

Made some tweaks to the planner's goal setting to take into account the last 5 messages as that additional context makes interacting with the planner less robot-like.

Finally, since I ended up simplifying the planner quite a bit I figured out that its possible for the planner to not have achieved its goal, but not explicitly succeeded either, so I decided to get cheeky and plumb the log into the planner to see if we can't make it a little more self-aware about when it gets things wrong; hopefully that'll make poking at the chat bot to "try again and think harder" actually work.

Also, made it so that we will start out with some known context with respect to what tools are available.