Add Zip file content to RAG; refactoring AddDirectoryToVectorStore and AddFileToVectorStore as a means of simplifying/unifying those flows.
Cleaned-up some of the RAG commands/menu.

Added supported file types to initial context (via RAG), and cleaned-up available tools context a bit while I was at it.

Created SearchVectorStore tool to make it so that the provider is capable of explicitly searching its knowledge (if it so chooses to -- un-tested/validated at the moment).

Some additional quality of life changes to RenderMenu (now with Home/End).

Standardized on screaming text at the LLM w.r.t. ParsedTypes.