Introduced MMR for better/smarter search results (can be toggled on/off).
Added GetEmbeddingsAsync to the IEmbeddingProvider interface with implementations for Ollama and AzureAI.  This should have a significant impact on embedding perf as we're batching these things up now instead of one-at-a-timing it.

Also took a stab at reducing the amount of text being sent to the LLM in the planner.